+++ src/m365/outlook/Outlook.ts
@@ -1,0 +1,21 @@
+export class Outlook {
+  public static wellKnownFolderNames: string[] = [
+    'archive',
+    'clutter',
+    'conflicts',
+    'conversationhistory',
+    'deleteditems',
+    'drafts',
+    'inbox',
+    'junkemail',
+    'localfailures',
+    'msgfolderroot',
+    'outbox',
+    'recoverableitemsdeletions',
+    'scheduled',
+    'searchfolders',
+    'sentitems',
+    'serverfailures',
+    'syncissues'
+  ];
+}