+++ src/m365/tenant/commands.ts
@@ -6,5 +6,6 @@ export default {
   TENANT_REPORT_ACTIVEUSERDETAIL: `${prefix} report activeuserdetail`,
   TENANT_REPORT_SERVICESUSERCOUNTS: `${prefix} report servicesusercounts`,
   TENANT_SERVICE_LIST: `${prefix} service list`,
+  TENANT_SERVICE_MESSAGE_LIST: `${prefix} service message list`,
   TENANT_STATUS_LIST: `${prefix} status list`
 };